+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Added logic to get a list of enabled modules in target Magento instance
+* Changed module file resolver to only search files in modules that are enabled in target Magento instance
+* Added dynamic generation of page classes based on enabled modules in target Magento instance
 * Changed path to generated page classes
 * Implemented recursive re-initialization for Block
 * Implemented scenario configuration reader